Revert "chore: rename and move filter options component"
This reverts commit f198d0d25558d2318700c9facab4ee4c5b1a7bca.
