+++ src/screens/search_results/EmptySearchResults.tsx
@@ -4,7 +4,7 @@ import { ShowCarousel, Button } from '../../components';
 import { useTrendingShows } from '../../hooks';
 import Home from '@mui/icons-material/Home';
 import { useNavigate } from 'react-router-dom';
-import SortFilterHeader from '../../components/SortFilterHeader';
+import SearchResultsHeader from './SearchResultsHeader';
 
 interface EmptySearchResultsProps {
     /**
