+++ src/screens/search_results/index.ts
@@ -1,6 +1,6 @@
 import EmptySearchResults from './EmptySearchResults';
 import SearchResultsCards from './SearchResultsCards';
-import SearchResultsHeader from '../../components/SortFilterHeader';
+import SearchResultsHeader from './SearchResultsHeader';
 import SearchResultsScreen from './SearchResultsScreen';
 
 export { EmptySearchResults, SearchResultsCards, SearchResultsHeader, SearchResultsScreen };
